Move all price logic to an utils
