+++ src/custom/state/lists/actions/actionsMod.ts
@@ -1,6 +1,7 @@
 import { ActionCreatorWithPayload, createAction } from '@reduxjs/toolkit'
 import { TokenList, Version } from '@uniswap/token-lists'
 import { SupportedChainId as ChainId } from 'constants/chains'
+export { SupportedChainId as ChainId } from 'constants/chains'
 
 export interface WithChainId {
   chainId?: ChainId
